Convert emf.js to typescript
