+++ package.json
@@ -8,7 +8,7 @@
         "fix": "eclint fix src/* src/**/* test/*.js",
         "clean": "rimraf build && rimraf dist",
         "compile": "tsc",
-        "bundle": "rollup --config rollup.config.rtfjs.js && rollup --config rollup.config.wmfjs.js",
+        "bundle": "rollup --config rollup.config.rtfjs.js && rollup --config rollup.config.wmfjs.js && rollup --config rollup.config.emfjs.js",
         "build": "npm run clean && npm run compile && npm run bundle",
         "generate-testcase": "node test/generate-testcase.js",
         "test": "mocha test/test-rtfjs.js"
