+++ samples/01_rtf/rtf.html
@@ -10,7 +10,7 @@
 
     <script src="../../dist/rtf.bundle.js"></script>
     <script src="../../dist/wmf.bundle.js"></script>
-    <script src="../../dist/emf.js"></script>
+    <script src="../../dist/emf.bundle.js"></script>
 
     <link href="../.common/css/jquery.svg.css" rel="stylesheet" />
