+++ samples/03_emf/emf.html
@@ -8,7 +8,7 @@
     <script src="../.common/dep/jquery.svg.min.js"></script>
     <script src="../.common/dep/jquery.svgfilter.min.js"></script>
 
-    <script src="../../dist/emf.js"></script>
+    <script src="../../dist/emf.bundle.js"></script>
 
     <link href="../.common/css/jquery.svg.css" rel="stylesheet" />
